add uint8 casting

+++ collect_false_positives.py
@@ -34,7 +34,7 @@ def read_input(readers, t0, a=None, tstep=1024, nchan=320):
     if a is None:
         a = np.zeros((nbeams, tstep, nchan, 1), dtype=np.float32)
     for i in range(nbeams):
-        readers[i].read_data(t_start=t0, t_stop=t0+tstep)
+        readers[i].read_data(t_start=t0, t_stop=t0+tstep).astype('uint8').astype(np.float32)
         a[i, ..., 0] = readers[i].data.squeeze()
     return a
